requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
banal==1.0.6
beautifulsoup4==4.12.3
brotlipy==0.7.0
cachetools==5.3.3
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
contourpy==1.2.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
einops==0.8.0
filelock==3.15.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gdcm==1.1
gdown==5.2.0
greenlet==3.0.3
grpcio==1.64.1
huggingface-hub==0.23.4
hydra-core==1.3.2
idna @ file:///croot/idna_1666125576474/work
Jinja2==3.1.4
kiwisolver==1.4.5
lightning==2.2.5
lightning-utilities==0.11.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
monai==1.3.1
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
nibabel==5.2.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
omegaconf==2.3.0
packaging==24.1
pandas==2.2.2
Pillow @ file:///croot/pillow_1696580024257/work
protobuf==4.25.3
psutil==5.9.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydicom==2.4.4
Pygments==2.18.0
pylibjpeg==2.0.0
pylibjpeg-libjpeg==2.1.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.2
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytz==2024.1
PyYAML==6.0.1
requests @ file:///croot/requests_1690400202158/work
rich==13.7.1
safetensors==0.4.3
SimpleITK==2.3.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.5
SQLAlchemy==1.4.53
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==1.0.7
torch==2.1.0+cu121
torch-tb-profiler==0.4.3
torchaudio==2.1.0+cu121
torchmetrics==1.4.0.post0
torchvision==0.16.0+cu121
tqdm==4.66.4
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1686163155763/work
Werkzeug==3.0.3
yarl==1.9.4
